Dear Professor,
you will find hereby my assignment6

I am very sorry for the delay, but I encountered of lot of difficulties
in doing this assignment.

In particular, I have trouble in setting my regular expression to test
the validity of the user input for question 5 so my program cannot
interact with the user as asked in question 5.

Nevertheless I have completed all other questions from 1 to 4 and it
works normally.

I will try to see if I can find someone to help me fix the problem with
the regular expression in Q5 tomorrow.

Best,

Vincent